rename: move FormTextInput to common folder
