+++ src/components/index.ts
@@ -36,7 +36,15 @@ export { default as CategoryChip } from './seller/common/CategoryChip';
 export { default as RadioInputSelector } from './seller/common/RadioInputSelector';
 export { default as SellerModal } from './seller/common/SellerModal';
 export { default as RadioBottomSheet } from './seller/common/RadioBottomSheet';
-
+export {
+  FormWideTextArea,
+  FormLimitedTextInput,
+  FormLinkInput,
+  FormPriceInput,
+  FormSalePriceInput,
+  FormEmailInput,
+  FormSNSInput,
+} from './common/FormTextInput';
 //item
 export { CategoryMultiSelector } from './seller/item/registration/CategoryMultiSelector';
 export {
@@ -67,12 +75,3 @@ export {
   DateTimePicker,
 } from './seller/item/registration/DateTimePicker';
 export { ItemSection } from './seller/item/registration/ItemSection';
-export {
-  FormWideTextArea,
-  FormLimitedTextInput,
-  FormLinkInput,
-  FormPriceInput,
-  FormSalePriceInput,
-  FormEmailInput,
-  FormSNSInput,
-} from './seller/item/registration/FormTextInput';
